Patch 1011
・脆弱性強化（clif.c, memorial.c, script.c）
・mob_talk() の引数をunsigned shortからintに変更（mob.*）

+++ src/map/mob.c
@@ -3734,7 +3734,7 @@ struct block_list * mob_selecttarget(struct mob_data *md, int range)
  * Mobの発言
  *------------------------------------------
  */
-int mob_talk(struct mob_data *md, unsigned short msg_id)
+int mob_talk(struct mob_data *md, int msg_id)
 {
 	char output[256];
